Revert "Fix tests for OpenSearch and blocked index/cluster after watermark hit (#15420)"
This reverts commit 259fffba9f7912eaaa8ec76c2932a8fb952c591f.
